Work Effects, Following Camera, Colliding with Monsters
-Effects
Effect that occurs when Giant Slime lands on the Ground.
Effect that occurs when Slime Fires.
Effect that occurs when Slime and Enemys are hit.
Effect that occurs when Slime get a Giant item.

-Follow Camera
Adjusting the distance according to the scale.
Camera Shaking.

-Colliding with Monsters
Ingore Colliding with Monsters when Slime is on Giant State.
Slime Jumps when it is above a Enemy.